requirements.txt
asgiref==3.2.10
astroid==2.4.2
bahttext==1.0.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.8
django-appconf==1.0.4
django-import-export==2.2.0
django-select2==7.4.2
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.4
psycopg2-binary==2.8.5
pylint==2.5.3
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
